# Conflicts:
#	src/main/java/org/example/Model/Data.json
#	src/main/java/org/example/Model/SoldierType.java

+++ src/main/java/org/example/Model/Soldier.java
@@ -4,8 +4,8 @@ import java.util.regex.Matcher;
 
 public class Soldier extends Person{
     int attackPower;
-    int defensePower;
-    int speed;
+    double defensePower;
+    double speed;
     int range;
     int secondRange = 0;
     int cost;
